+++ AppLovin/source/plugin/Assets/GoogleMobileAds/Mediation/AppLovin/Common/DummyClient.cs
@@ -24,17 +24,7 @@ namespace GoogleMobileAds.Mediation.AppLovin.Common
             Debug.Log("Dummy " + MethodBase.GetCurrentMethod().Name);
         }
 
-        public void Initialize()
-        {
-            Debug.Log("Dummy " + MethodBase.GetCurrentMethod().Name);
-        }
-
         public void SetHasUserConsent(bool hasUserConsent)
-        {
-            Debug.Log("Dummy " + MethodBase.GetCurrentMethod().Name);
-        }
-
-        public void SetIsAgeRestrictedUser(bool isAgeRestrictedUser)
         {
             Debug.Log("Dummy " + MethodBase.GetCurrentMethod().Name);
         }