+++ AppLovin/source/plugin/Assets/GoogleMobileAds/Mediation/AppLovin/Common/IAppLovinClient.cs
@@ -16,9 +16,7 @@ namespace GoogleMobileAds.Mediation.AppLovin.Common
 {
     public interface IAppLovinClient
     {
-        void Initialize();
         void SetHasUserConsent(bool hasUserConsent);
-        void SetIsAgeRestrictedUser(bool isAgeRestrictedUser);
         void SetDoNotSell(bool doNotSell);
     }
 }