+++ AppLovin/source/plugin/Assets/GoogleMobileAds/Mediation/AppLovin/Plugins/iOS/GADUMAppLovinInterface.m
@@ -14,14 +14,8 @@
 
 #import <AppLovinSDK/AppLovinSDK.h>
 
-void GADUMInitializeAppLovin(void) { [ALSdk initializeSdk]; }
-
 void GADUMAppLovinSetHasUserConsent(BOOL hasUserConsent) {
   [ALPrivacySettings setHasUserConsent:hasUserConsent];
-}
-
-void GADUMAppLovinSetIsAgeRestrictedUser(bool isAgeRestrictedUser) {
-  [ALPrivacySettings setIsAgeRestrictedUser:isAgeRestrictedUser];
 }
 
 void GADUMAppLovinSetDoNotSell(bool doNotSell) { [ALPrivacySettings setDoNotSell:doNotSell]; }